backend/Hasity/requirements.txt
channels==4.0.0
Django==4.0.6
djangorestframework==3.14.0
openai==0.25.0
PyMySQL==1.0.2
